OverlayWindow.[hm]: fix a type warning, and move an #import to the .m

+++ Classes/Overlay Widgets/OverlayWindow.m
@@ -41,10 +41,11 @@
  */
 
 #import "OverlayWindow.h"
+#import "NiceWindow.h"
 
 @implementation OverlayWindow
 
--(id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
+-(id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
 {
     
     if((self = [super initWithContentRect:contentRect
